+++ apps/experiments/tests/test_views.py
@@ -23,7 +23,7 @@ from apps.experiments.models import (
 )
 from apps.experiments.views.experiment import (
     ExperimentForm,
-    ExperimentTableView,
+    ExperimentExperimentTableView,
     _verify_user_or_start_session,
 )
 from apps.teams.backends import add_user_to_team
@@ -439,7 +439,7 @@ class TestExperimentTableView:
 
         request = RequestFactory().get(reverse("experiments:table", args=[team.slug]))
         request.team = team
-        view = ExperimentTableView()
+        view = ExperimentExperimentTableView()
         view.request = request
         assert list(view.get_queryset().all()) == [experiment]
