+++ apps/experiments/urls.py
@@ -40,7 +40,7 @@ urlpatterns = [
     ),
     # experiments
     path("new/", views.CreateExperiment.as_view(), name="new"),
-    path("table/", views.ExperimentTableView.as_view(), name="table"),
+    path("table/", views.ExperimentExperimentTableView.as_view(), name="table"),
     path("e/<int:experiment_id>/", views.single_experiment_home, name="single_experiment_home"),
     path("e/<int:experiment_id>/sessions-table/", views.ExperimentSessionsTableView.as_view(), name="sessions-list"),
     path("e/<int:experiment_id>/versions/", views.ExperimentVersionsTableView.as_view(), name="versions-list"),
